perf(refactor address validation to have 1 file instead of 3): refactor address validation to have 1
refactor address validation to have 1 file instead of 3
