+++ packages/forms-web-app/src/routes/register/myself/address.js
@@ -2,7 +2,7 @@ const express = require('express');
 
 const addressController = require('../../../controllers/register/myself/address');
 const { validationErrorHandler } = require('../../../validators/validation-error-handler');
-const { rules: addressValidationRules } = require('../../../validators/register/myself/address');
+const { addressValidationRules } = require('../../../validators/shared');
 
 const router = express.Router();
