+++ packages/forms-web-app/src/validators/shared/index.js
@@ -1,7 +1,9 @@
 const { validateNotEmpty } = require('./not-empty');
 const { validateNotEmptyAndLength } = require('./not-empty-and-length');
+const { addressValidationRules } = require('./address');
 
 module.exports = {
 	validateNotEmpty,
-	validateNotEmptyAndLength
+	validateNotEmptyAndLength,
+	addressValidationRules
 };
